Isolate dependencies so that only XGPSSDK folder can be copied to project
