+++ XGPSSample/TripViewController.swift
@@ -17,19 +17,6 @@ class TripsCell : UITableViewCell {
     
 }
 
-struct LogData {
-    let sig: Int
-    let interval: Int
-    let startBlock: Int
-    let countEntry: Int
-    let countBlock: Int
-    let createDate: String
-    let createTime: String
-    let fileSize: Int
-    let defaultString: String
-    var localFilename: String
-}
-
 class TripViewController: UITableViewController, TripLogDelegate {
     @IBOutlet weak var spinner: UIActivityIndicatorView!
 //    @IBOutlet weak var topTitleBar: UINavigationItem!
