+++ XGPSSample/XGPSSDK/LogData.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+struct LogData {
+    let sig: Int
+    let interval: Int
+    let startBlock: Int
+    let countEntry: Int
+    let countBlock: Int
+    let createDate: String
+    let createTime: String
+    let fileSize: Int
+    let defaultString: String
+    var localFilename: String
+}
+
